alignment/align_undl_text/requirements_client.txt
argostranslate
requests[socks]
brotli
PySocks